refactor: introduce RelationalConfigBase for shared database configuration validation
